Completed delete button and gave textarea maxlength

+++ src/index.js
@@ -27,18 +27,22 @@ const checkIfId = () => {
 };
 
 const deleteItem = (id) => {
-  const idArr = [];
-  for (let i = 0; i < parseInt(idGet(), 10) - 1; i += 1) {
-    idArr.push(JSON.parse(localStorage.getItem(i.toString())));
-    idArr[i].hisId = i;
+  if (parseInt(idGet(), 10) !== 0) {
+    const idArr = [];
+    for (let i = 0; i < parseInt(idGet(), 10); i += 1) {
+      idArr.push(JSON.parse(localStorage.getItem(i.toString())));
+      idArr[i].hisId = i;
+    }
+    const newId = parseInt(idGet(), 10) - 1;
+    localStorage.clear();
+    console.log(idArr);
+    idArr.splice(parseInt(id, 10), 1);
+    console.log(idArr);
+    for (let i = 0; i < idArr.length; i += 1) {
+      localStorage.setItem(i.toString(), JSON.stringify(idArr[i]));
+    }
+    localStorage.setItem("Id", newId.toString());
   }
-  const newId = parseInt(idGet(), 10) - 1;
-  localStorage.clear();
-  idArr.splice(parseInt(id, 10), 1);
-  for (let i = 0; i < idArr.length; i += 1) {
-    localStorage.setItem(i.toString(), idArr[i]);
-  }
-  localStorage.setItem("Id", newId.toString());
 };
 
 const requestItem = (id, item) => {
@@ -150,6 +154,7 @@ const checkPage = (value) => {
   let submitBtns;
   let writeBtns;
   let nextBtn;
+  let deleteBtn;
   let prevBtn;
   const pageBtns = {};
   if (currentPage === "main") {
@@ -174,10 +179,19 @@ const checkPage = (value) => {
   } else if (currentPage === "see") {
     btnLogic = value.children["0"].children["0"];
     nextBtn = value.children["0"].children["2"].children["0"];
-    prevBtn = value.children["0"].children["3"].children["0"];
+    deleteBtn = value.children["0"].children["3"].children["0"];
+    prevBtn = value.children["0"].children["4"].children["0"];
   }
 
-  return { btnLogic, pageBtns, submitBtns, writeBtns, nextBtn, prevBtn };
+  return {
+    btnLogic,
+    pageBtns,
+    submitBtns,
+    writeBtns,
+    nextBtn,
+    deleteBtn,
+    prevBtn,
+  };
 };
 
 //
@@ -200,7 +214,9 @@ const SvgPict = () => {
   const writeIcon = `<svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" x="0" y="0" viewBox="0 0 256 256" style="enable-background:new 0 0 256 256" xml:space="preserve"><style>.st2{fill:#69ebfc}.st5{fill:#d476e2}.st14{fill:#edd4c2}</style><path d="M205 0H20c5.52 0 10.52 2.24 14.14 5.86S40 14.48 40 20v236h185V20c0-11.046-8.954-20-20-20z" style="fill:#ffefe4"/><path class="st14" d="M40 250h185v6H40z"/><path d="M205 0H20c5.52 0 10.52 2.24 14.14 5.86.045.045.083.095.127.14H205c11.046 0 20 8.954 20 20v-6c0-11.046-8.954-20-20-20z" style="fill:#fff7f2"/><path d="M144 56H70a8 8 0 0 1 0-16h74a8 8 0 0 1 0 16z" style="fill:#ff8354"/><path class="st14" d="M188 84H66a4 4 0 0 1 0-8h122a4 4 0 0 1 0 8zM188 111.2H66a4 4 0 0 1 0-8h122a4 4 0 0 1 0 8zM188 138.4H66a4 4 0 0 1 0-8h122a4 4 0 0 1 0 8zM188 165.6H66a4 4 0 0 1 0-8h122a4 4 0 0 1 0 8zM188 192.8H66a4 4 0 0 1 0-8h122a4 4 0 0 1 0 8zM147 220H66a4 4 0 0 1 0-8h81a4 4 0 0 1 0 8zM40 20v60H0V20C0 8.95 8.95 0 20 0c5.52 0 10.52 2.24 14.14 5.86S40 14.48 40 20z"/><path transform="rotate(30 173.417 175.565)" class="st5" d="M163.411 150.575h20v50h-20z"/><path class="st5" d="m169.571 202.226-17.32-10-5.625 24.567a2.708 2.708 0 0 0 4.483 2.589l18.462-17.156z"/><path transform="rotate(30 212.163 108.447)" class="st2" d="M202.161 55.958h20v105h-20z"/><path transform="rotate(-150 244.807 75.92)" class="st2" d="M242.803 72.92h4v6h-4z"/><path class="st2" d="M223.999 123.954a4 4 0 0 0 5.464-1.464l26-45.033a4 4 0 1 0-6.928-4l-26 45.033a4 4 0 0 0 1.464 5.464z"/><path class="st5" d="M250.911 41.341c-4.783-2.761-10.899-1.123-13.66 3.66l-7.5 12.99 17.32 10 7.5-12.99c2.762-4.782 1.123-10.898-3.66-13.66z"> </path></svg>`;
 
   const noteIcon = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><title>note-multiple-outline</title><path d="M3,6V22H21V24H3A2,2 0 0,1 1,22V6H3M16,9H21.5L16,3.5V9M7,2H17L23,8V18A2,2 0 0,1 21,20H7C5.89,20 5,19.1 5,18V4A2,2 0 0,1 7,2M7,4V18H21V11H14V4H7Z"> </path></svg>`;
-  return { plusIcon, returnIcon, writeIcon, noteIcon };
+
+  const deleteIcon = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><title>delete-forever-outline</title><path d="M14.12,10.47L12,12.59L9.87,10.47L8.46,11.88L10.59,14L8.47,16.12L9.88,17.53L12,15.41L14.12,17.53L15.53,16.12L13.41,14L15.53,11.88L14.12,10.47M15.5,4L14.5,3H9.5L8.5,4H5V6H19V4H15.5M6,19A2,2 0 0,0 8,21H16A2,2 0 0,0 18,19V7H6V19M8,9H16V19H8V9Z" /></svg>`;
+  return { deleteIcon, plusIcon, returnIcon, writeIcon, noteIcon };
 };
 
 const cardsInner = {
@@ -288,6 +304,7 @@ const TodoText = (pageName) => {
     "",
     `${pageName}-projText`
   );
+  TodoTextInside.setAttribute("maxlength", "100");
 
   TodoTextContainer.append(TodoTextInside);
 
@@ -365,6 +382,13 @@ function SeePage(id, value) {
   );
 
   this.next = methods.methodCreate("button", "next", "btn-next");
+  this.delete = methods.methodCreate(
+    "button",
+    "",
+    "btn-delete",
+    "delete",
+    SvgPict().deleteIcon
+  );
   this.prev = methods.methodCreate("button", "previus", "btn-prev");
 
   this.name = methods.divCreate("Name:", "item-name");
@@ -379,15 +403,18 @@ function SeePage(id, value) {
   this.field.append(this.name, this.time, this.date, this.text);
 
   this.nextContainer = methods.divCreate("", "btn-next-container");
+  this.deleteContainer = methods.divCreate("", "btn-delete-container");
   this.prevContainer = methods.divCreate("", "btn-prev-container");
 
   this.nextContainer.append(this.next);
+  this.deleteContainer.append(this.delete);
   this.prevContainer.append(this.prev);
 
   this.page.append(
     this.returnBtn,
     this.field,
     this.nextContainer,
+    this.deleteContainer,
     this.prevContainer
   );
 }
@@ -567,6 +594,7 @@ const seeBtnsLogic = (value, id, page) => {
   const field = value.children["0"];
   const next = checkPage(value).nextBtn;
   const prev = checkPage(value).prevBtn;
+  const deleter = checkPage(value).deleteBtn;
   let currId = id;
   const arrUsable = EleNum(page);
   next.addEventListener("click", () => {
@@ -603,6 +631,19 @@ const seeBtnsLogic = (value, id, page) => {
         seePage(arrUsable[currId], "see", page).myField,
         field.children["1"]
       );
+    }
+  });
+  deleter.addEventListener("click", () => {
+    if (parseInt(idGet(), 10) !== 0) {
+      const confirms = confirm("Do you really want to delete this note");
+      if (confirms) {
+        deleteItem(currId);
+        field.removeChild(field.children["1"]);
+        field.insertBefore(
+          seePage(arrUsable[currId], "see", page).myField,
+          field.children["1"]
+        );
+      }
     }
   });
 };
